+++ test/platform_handler_test.dart
@@ -1,3 +1,1 @@
-import 'package:platform_handler/platform_handler.dart';
-
 void main() {}